Better data fetching strategy
Unlike CoRT, Sentinel doesn't use notifications, so it's waste of
resources to call the API every minute if the page isn't used. The
ticker option is not ideal here indeed. Instead, work on main thread
and add some rules:

Data is refreshed :
	- immediatly, when page is loaded
	- every minutes between :10 and :15 if the window is visible
	- immediatly, when the page is focused

Data is not refreshed :
 	- if the window is hidden
 	- if the last refresh was less than 5s ago (avoid focus event
	  spam if hopping between tabs)

sw.js was removed since it's only needed if notifications are used.